Revert "[chttp2] Improve huffman decode efficiency (#30479)" (#30961)
This reverts commit 6c7f921f5f13763d6e09296825cdb99f9e000115.

+++ src/core/ext/transport/chttp2/transport/bin_encoder.cc
@@ -98,7 +98,7 @@ grpc_slice grpc_chttp2_huffman_compress(const grpc_slice& input) {
   const uint8_t* in;
   uint8_t* out;
   grpc_slice output;
-  uint64_t temp = 0;
+  uint32_t temp = 0;
   uint32_t temp_length = 0;
 
   nbits = 0;
